backend/requirements.txt
Django>=5.0
djangorestframework
psycopg2-binary
dj-database-url
google-generativeai